requirements.txt
python-telegram-bot==13.15
requests==2.31.0
matplotlib==3.7.1
